backend/requirements.txt
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
django-cors-headers==3.13.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
mixer==7.2.2
graphene-django==3.0.0
psycopg2==2.9.5